+++ quant_strategy_tokenizer/hash_v2/__init__.py
@@ -3,10 +3,12 @@
 from quant_strategy_tokenizer.hash_v2.audit_chain_hash import audit_chain_hash_v2
 from quant_strategy_tokenizer.hash_v2.behavior_hash import behavior_hash_v2
 from quant_strategy_tokenizer.hash_v2.common import HASH_V2_PATTERN, IRHashesV2, hash_v2_payload
+from quant_strategy_tokenizer.hash_v2.expected_artifact_hash import expected_artifact_hash_v2
 from quant_strategy_tokenizer.hash_v2.graph_hash import graph_hash_v2
 from quant_strategy_tokenizer.hash_v2.implementation_ref_hash import implementation_ref_hash_v2
 from quant_strategy_tokenizer.hash_v2.instance_hash import compute_hashes_v2, instance_hash_v2
 from quant_strategy_tokenizer.hash_v2.param_hash import param_hash_v2
+from quant_strategy_tokenizer.hash_v2.runtime_environment_hash import runtime_environment_hash_v2
 from quant_strategy_tokenizer.hash_v2.signature_hash import (
     signature_hash_for_ports_v2,
     signature_hash_v2,
@@ -20,11 +22,13 @@ __all__ = [
     "audit_chain_hash_v2",
     "behavior_hash_v2",
     "compute_hashes_v2",
+    "expected_artifact_hash_v2",
     "graph_hash_v2",
     "hash_v2_payload",
     "implementation_ref_hash_v2",
     "instance_hash_v2",
     "param_hash_v2",
+    "runtime_environment_hash_v2",
     "signature_hash_for_ports_v2",
     "signature_hash_v2",
     "token_pack_hash_v2",