+++ quant_strategy_tokenizer/types_v2/__init__.py
@@ -1,9 +1,12 @@
 """Token System v2 structured type models."""
 
 from quant_strategy_tokenizer.types_v2.type_spec import (
+    TYPE_SPEC_SCHEMA_VERSION,
     AvailableAt,
     Clock,
     IntrinsicTemporalSpec,
+    MissingPolicy,
+    SelectionKind,
     TypeKind,
     TypeSpec,
     parse_type_spec,
@@ -11,9 +14,12 @@ from quant_strategy_tokenizer.types_v2.type_spec import (
 from quant_strategy_tokenizer.types_v2.value_type import ValueType, ValueTypeName
 
 __all__ = [
+    "TYPE_SPEC_SCHEMA_VERSION",
     "AvailableAt",
     "Clock",
     "IntrinsicTemporalSpec",
+    "MissingPolicy",
+    "SelectionKind",
     "TypeKind",
     "TypeSpec",
     "ValueType",